pos table of 32 bit size sent to the GPU to reduce GPU global memory used

+++ GPU.h
@@ -24,7 +24,7 @@ struct hsp {
 };
 
 typedef size_t(*InitializeProcessor_ptr)(int t, int f);
-typedef void(*SendSeedPosTable_ptr)(uint32_t* index_table, uint32_t index_table_size, uint64_t* pos_table, uint32_t ref_size);
+typedef void(*SendSeedPosTable_ptr)(uint32_t* index_table, uint32_t index_table_size, uint32_t* pos_table, uint32_t ref_size);
 typedef std::vector<hsp> (*SeedAndFilter_ptr)(std::vector<uint64_t> seed_offset_vector, bool rev);
 typedef void(*ShutdownProcessor_ptr)();
 typedef void(*SendRefWriteRequest_ptr)(size_t addr, size_t len);
